Former-commit-id: 8a9d876760bc4d4c4a75eba39209ce0e4931c8db
Former-commit-id: b4f063bd73d4ecd8fed1314c6068b643525d6165
Former-commit-id: 08928defeb8bab9dcfb146da4a33106181b7fc9f

+++ vizia_api_bindings/python/ViziaDoomGamePython.h
@@ -36,8 +36,8 @@ namespace Vizia {
         PythonState getState();
         boost::python::list getLastAction();
         object getGameScreen();
-        float makeAction(boost::python::list &action);
-        float makeAction(boost::python::list &action, unsigned int tics);
+        double makeAction(boost::python::list &action);
+        double makeAction(boost::python::list &action, unsigned int tics);
 
     private:
